move pieces to utility files, update tests based on it, and add more sanity checks
