+++ fairlearn/post_processing/_constants.py
@@ -1,0 +1,6 @@
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# Licensed under the MIT License.
+
+SCORE_KEY = "score"
+LABEL_KEY = "label"
+ATTRIBUTE_KEY = "attribute"
